// interfaces/b5.interface.ts
export interface IPagination {
    row: string | number
    page: string | number
    key: string
    period: string
    limit: number | string
}